+++ src/test/java/stage2/CalculatorTest.java
@@ -49,69 +49,4 @@ public class CalculatorTest {
 
         assertThat(throwable).isExactlyInstanceOf(IllegalArgumentException.class);
     }
-
-    @Test
-    @DisplayName("사칙연산 기능 테스트 - 덧셈")
-    void calculatorAddTest() {
-        String input = "1 + 2 + 3 + 4 + 5";
-        Integer expectedValue = 15;
-
-        Integer result = c.calculate(input);
-
-        assertThat(result).isEqualTo(expectedValue);
-    }
-
-    @Test
-    @DisplayName("사칙연산 기능 테스트 - 뺄셈")
-    void calculatorMinusTest() {
-        String input = "50 - 10 - 5 - 1";
-        Integer expectedValue = 34;
-
-        Integer result = c.calculate(input);
-
-        assertThat(result).isEqualTo(expectedValue);
-    }
-
-    @Test
-    @DisplayName("사칙연산 기능 테스트 - 곱셈")
-    void calculatorMultiTest() {
-        String input = "2 * 2 * 2 * 2";
-        Integer expectedValue = 16;
-
-        Integer result = c.calculate(input);
-
-        assertThat(result).isEqualTo(expectedValue);
-    }
-
-    @Test
-    @DisplayName("사칙연산 기능 테스트 - 나눗셈")
-    void calculatorDivideTest() {
-        String input = "50 / 5 / 5 / 2";
-        Integer expectedValue = 1;
-
-        Integer result = c.calculate(input);
-
-        assertThat(result).isEqualTo(expectedValue);
-    }
-
-    @Test
-    @DisplayName("사칙연산 기능 테스트 - 나눗셈 0")
-    void calculatorDivideInvalidTest() {
-        String input = "10 / 0";
-
-        Throwable throwable = catchThrowable(() -> c.calculate(input));
-
-        assertThat(throwable).isExactlyInstanceOf(ArithmeticException.class);
-
-    }
-
-    @Test
-    @DisplayName("잘못된 연산자 넣었을 경우 테스트")
-    void calculatorInvalidOperationTest() {
-        String input = "1 ? 2";
-
-        Throwable throwable = catchThrowable(() -> c.calculate(input));
-
-        assertThat(throwable).isExactlyInstanceOf(IllegalArgumentException.class);
-    }
 }
